smart-admin-ai/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
